requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cinemagoer==2022.2.11
Django==4.0.2
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
lxml==4.7.1
pytz==2021.3
requests==2.27.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
urllib3==1.26.8
